moved EXPORT_NOTEBOOK logic into exportNotebook action creator
